refactor Discord webhook price update
