+++ src/lib/DiscordWebhook/export.ts
@@ -2,7 +2,16 @@ import sendAlert from './sendAlert';
 import sendOfferReview from './sendOfferReview';
 import sendTradeSummary from './sendTradeSummary';
 import sendPartnerMessage from './sendPartnerMessage';
+import sendWebHookPriceUpdateV1 from './pricelistUpdate';
 import { Webhook } from './interfaces';
 import { sendWebhook } from './utils';
 
-export { sendAlert, sendOfferReview, sendTradeSummary, sendPartnerMessage, Webhook, sendWebhook };
+export {
+    sendAlert,
+    sendOfferReview,
+    sendTradeSummary,
+    sendPartnerMessage,
+    sendWebHookPriceUpdateV1,
+    Webhook,
+    sendWebhook
+};
